requirements.txt
yt-dlp>=2024.1.0
ffmpeg-python>=0.2.0
click>=8.0.0
colorama>=0.4.4
